+++ mangadex_downloader/cli/__init__.py
@@ -22,8 +22,10 @@ _deprecated_opts = {
     # But i need the warning feature, hehe
     "range": "--range is disabled, because it's broken and need to rework",
 
-    "no_progress_bar": "--no-progress bar is deprecated, use '--progress-bar-layout=none' instead",
-    "verbose": "--verbose is deprecated, use '--log-level=DEBUG' instead",
+    "no_progress_bar": "--no-progress bar is deprecated and will be removed in v3.0.0" \
+                       ". Use '--progress-bar-layout=none' instead",
+    "verbose": "--verbose is deprecated and will be removed in v3.0.0. " \
+               "Use '--log-level=DEBUG' instead",
 }
 
 def check_deprecated_options(log, args):